Update
Updated Comments and organized all scripts. Removed redundant and unused code/scripts and methods.
Renamed scripts to match what they are doing since changes discussed with Tom, Scott and Jemma.

Merged CandyManager, EffectsManager and RoomManager into one script now called GameManager.

Work done with Tom-
Parented enemies to their spawners so they are removed when a room is deleted.

Health bar now increments during sugar rush and decrements otherwise by 1 second.

Changed item prefabs to an array and when an enemy is killed and an item is meant to drop, it will choose a random item from that array of item prefabs.

Basic Framework for PartyPopper ability added as a prefab.(Needs to be finished)

Fixed destroying of parent object for partypopper - Sarah

BUG: PumpkinCrawler is stretching weird when going over a table?
Co-Authored-By: GaiVaige <127663263+GaiVaige@users.noreply.github.com>