Update for Beta V2
-Player attack has been updated. Combo 2 and 3 have added with code and working animations.
-Adjusted collisions on enemies and player to be on the ground accurately.
-BUG FIX: Player was only toggling death screen when health reached 0 but not if the timer ran out.
-Player is now invulnerable during sugar rush.
-Player now deals double damage in sugar rush.
-Added sugar rush version of chattering skulls and drop implementation in Health script.

NOTE: Began implementing new party room. Enemies could walk through the big table. Make sure to check objects are on the ground for NavMesh to create accurate pathing.  - Also added a invisible collision box for the table, as even on the ground it creates pathing on top of the table and underneath.
coloured balloon prefabs did not exist and could not be moved.